Gestion-Boite-Maile

+++ User/src/main/java/com/example/user/Repository/IuserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface IuserRepository extends JpaRepository<User,Long> {
+public interface IuserRepository extends JpaRepository<User,Integer > {
     Optional<User> findByEmail(String email);
 }
